Add files via upload
In many fire-related incidents, delayed detection and reporting often lead to severe property damage, environmental degradation, and loss of life. Traditional fire alert systems are either manual or limited in coverage and response capability. There is a need for an intelligent, real-time system that can assess fire severity and autonomously determine the need for emergency response.